requirements.txt
ansible-tower-cli==3.2.0
boto==2.48.0
botocore==1.7.33
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
docutils==0.14
idna==2.6
jmespath==0.9.3
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
wheel==0.24.0